refactor: OAuthProvider 파일 디렉토리 수정(entity -> domain)
