+++ src/main/kotlin/com/ssak3/timeattack/member/domain/Member.kt
@@ -33,9 +33,3 @@ data class Member(
             )
     }
 }
-
-// OAuth Provider Enum
-enum class OAuthProvider {
-    KAKAO,
-    GOOGLE,
-}
