+++ src/main/kotlin/com/ssak3/timeattack/member/domain/OAuthProvider.kt
@@ -1,0 +1,6 @@
+package com.ssak3.timeattack.member.domain
+
+enum class OAuthProvider {
+    KAKAO,
+    GOOGLE,
+}
