+++ src/main/kotlin/com/ssak3/timeattack/member/service/AuthService.kt
@@ -7,8 +7,8 @@ import com.ssak3.timeattack.member.auth.oidc.OIDCPayload
 import com.ssak3.timeattack.member.auth.oidc.OIDCTokenVerification
 import com.ssak3.timeattack.member.controller.LoginRequest
 import com.ssak3.timeattack.member.domain.Member
+import com.ssak3.timeattack.member.domain.OAuthProvider
 import com.ssak3.timeattack.member.repository.MemberRepository
-import com.ssak3.timeattack.member.repository.entity.OAuthProvider
 import com.ssak3.timeattack.member.repository.entity.OAuthProviderInfo
 import org.springframework.stereotype.Service
